fix big chain ids

+++ packages/core/src/database/index.ts
@@ -695,7 +695,7 @@ CREATE TABLE IF NOT EXISTS "${namespace.schema}"."_ponder_meta" (
             sql.raw(`
 CREATE TABLE IF NOT EXISTS "${namespace.schema}"."_ponder_checkpoint" (
   "chain_name" TEXT PRIMARY KEY,
-  "chain_id" INTEGER NOT NULL,
+  "chain_id" BIGINT NOT NULL,
   "safe_checkpoint" VARCHAR(75) NOT NULL,
   "latest_checkpoint" VARCHAR(75) NOT NULL
 )`),
